requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
flask-mail==0.10.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
anthropic==0.34.2
stripe==10.8.0
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
pypdf2==3.0.1
pdfminer.six==20231228
requests==2.32.3
gunicorn==23.0.0
werkzeug==3.0.4
bcrypt==4.2.0
itsdangerous==2.2.0
wtforms==3.1.2
email-validator==2.2.0
